initial commit of account_db functionality to branch
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/FgpUtil/branches/account_db@20050 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Util/src/main/java/org/gusdb/fgputil/db/platform/DBPlatform.java
@@ -51,7 +51,7 @@ public abstract class DBPlatform {
     // platform-dependent abstract methods
     //#########################################################################
 
-    public abstract int getNextId(DataSource dataSource, String schema,
+    public abstract long getNextId(DataSource dataSource, String schema,
         String table) throws SQLException, DBStateException;
 
     public abstract String getNextIdSqlExpression(String schema, String table);
